Renamed test folder to test utils
